Add support for publishing debug releases
This PR adds CI functionality so that pushing a branch `debug-release/foobar`
will cause Docker images to be published with the tag `debug-foobar-$COMMIT`.
These published and tagged images can then be used to test e.g. a specific PR
on a real cluster instance.

Context:

As part of PR #2884 I wanted to test my change on a real, large cluster
instance of Sourcegraph.

The only way for me to get CI to publish the images needed for doing this
currently was to tag an actual release candidate like `v3.3.0-debug.0` but in
my mind this is wrong to do because that looks like it actually corresponds to
the version 3.3.0 when in reality my change is just arbitrary and not directly
a statement of releasing a version (as cutting a release candidate may be, for
example).