Make selections specific to a code view
Having selections as a property of `CodeHost` is a problem, because [it is incorrect to assume that all code views on the page have the same selections](https://github.com/sourcegraph/sourcegraph/pull/2909#discussion_r268261523).

This adds `getSelections` and `observeSelections as optional properties of `CodeViewSpec`, and changes the code to make sure that when selections change, we don't update _all visible view components_ with the same spec of selection.

This assumes that only adding `getSelections()` and `observeSelections()` to single file code views is "good enough", which is still an imperfect assumption, but at least avoids having code that explicitly sets the same selections on a bunch of code views.